Final Commit(Continues on main rep)

+++ src/special/Evaluation.java
@@ -7,9 +7,9 @@ import special.Models.Lecture;
 public class Evaluation {
 
     public void Decider(List<Lecture> LectList, List<Metric> MetricList){
-        List<Integer> resultList = new ArrayList<Integer>();
+        List<Double> resultList = new ArrayList<Double>();
         for(Metric metric : MetricList){
-            int score = metric.evaluate(LectList);
+            double score = metric.evaluate(LectList);
             resultList.add(score);
         }
